tratamento de exceção update e delete

+++ src/main/java/com/clientedb/Clientedatabase/services/ClientService.java
@@ -11,24 +11,26 @@ import com.clientedb.Clientedatabase.entities.Client;
 import com.clientedb.Clientedatabase.repositories.ClientRepository;
 import com.clientedb.Clientedatabase.services.exceptions.ResourceNotFoundException;
 
+import jakarta.persistence.EntityNotFoundException;
+
 @Service
 public class ClientService {
-	
+
 	@Autowired
 	private ClientRepository repo;
-	
+
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
 		Client result = repo.findById(id).orElseThrow(() -> (new ResourceNotFoundException("Id inválido")));
 		return new ClientDTO(result);
 	}
-	
+
 	@Transactional(readOnly = true)
-	public Page<ClientDTO> findAll(Pageable pageable){
+	public Page<ClientDTO> findAll(Pageable pageable) {
 		Page<Client> dto = repo.findAll(pageable);
 		return dto.map(x -> new ClientDTO(x));
 	}
-	
+
 	@Transactional
 	public ClientDTO insert(ClientDTO dto) {
 		Client client = new Client();
@@ -40,22 +42,30 @@ public class ClientService {
 		client = repo.save(client);
 		return new ClientDTO(client);
 	}
-	
+
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
-		Client client = repo.getReferenceById(id);
-		client.setName(dto.getName());
-		client.setCpf(dto.getCpf());
-		client.setIncome(dto.getIncome());
-		client.setBirthDate(dto.getBirthDate());
-		client.setChildren(dto.getChildren());
-		client = repo.save(client);
-		return new ClientDTO(client);
+		try {
+			Client client = repo.getReferenceById(id);
+			client.setName(dto.getName());
+			client.setCpf(dto.getCpf());
+			client.setIncome(dto.getIncome());
+			client.setBirthDate(dto.getBirthDate());
+			client.setChildren(dto.getChildren());
+			client = repo.save(client);
+			return new ClientDTO(client);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Id inválido");
+		}
 	}
-	
+
 	@Transactional
 	public void delete(Long id) {
+		if (!repo.existsById(id)) {
+			throw new ResourceNotFoundException("Id inválido");
+		}
 		repo.deleteById(id);
-		
+
 	}
+
 }
